增强ISpanSerializable支持及序列化性能优化
本次提交重构了数据包序列化流程，统一并优化了 ISpanSerializable 类型的处理方式。DbTable、DefaultPacketEncoder、HttpResponse、StandardCodec 等核心组件均集成了高效的序列化/反序列化逻辑。ECKey 密钥类实现了 ISpanSerializable 接口，SpanSerializer 新增了 IPacket 泛型反序列化方法。同步修正了相关测试用例，提升类型安全与资源管理能力。整体提升了系统对高性能序列化场景的支持。

+++ XUnitTest.Core/Buffers/SpanWriterTests.cs
@@ -252,7 +252,7 @@ public class SpanWriterTests
         Assert.Equal(0, ms.Length);
 
         // 再写 8 字节，超出缓冲区 → 触发自动 Flush + 继续写入
-        writer.Write((Int64)0x5566778899AABBCC);
+        writer.Write(0x5566778899AABBCC);
         Assert.True(ms.Length > 0);
 
         writer.Flush();
